版本发布

+++ dangjia-group/dangjia-goods/dangjia-api-goods/src/main/java/com/dangjia/acg/dto/actuary/FlowDTO.java
@@ -2,6 +2,7 @@ package com.dangjia.acg.dto.actuary;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -13,6 +14,6 @@ import java.util.List;
 public class FlowDTO {
     private String name;
     private int type;//type 人工1 材料2 服务3
-    private Double sumTotal;//合计
+    private BigDecimal sumTotal;//合计
     private List<FlowActuaryDTO> flowActuaryDTOList;
 }
